use size_t whenever possible, only polOrder is int

+++ dune/gdt/assembler/local/codim0.hh
@@ -68,7 +68,8 @@ public:
    * type of systemMatrix
    *  \tparam R           RangeFieldType, i.e. double
    */
-  template <class T, int Td, int Tr, int TrC, class A, int Ad, int Ar, int ArC, class EntityType, class M, class R>
+  template <class T, size_t Td, size_t Tr, size_t TrC, class A, size_t Ad, size_t Ar, size_t ArC, class EntityType,
+            class M, class R>
   void assembleLocal(const SpaceInterface<T, Td, Tr, TrC>& testSpace, const SpaceInterface<A, Ad, Ar, ArC>& ansatzSpace,
                      const EntityType& entity, Dune::Stuff::LA::MatrixInterface<M, R>& systemMatrix,
                      std::vector<std::vector<Dune::DynamicMatrix<R>>>& tmpLocalMatricesContainer,
@@ -149,7 +150,7 @@ public:
    * type of systemVector
    *  \tparam R           RangeFieldType, i.e. double
    */
-  template <class T, int d, int r, int rC, class EntityType, class V, class R>
+  template <class T, size_t d, size_t r, size_t rC, class EntityType, class V, class R>
   void assembleLocal(const SpaceInterface<T, d, r, rC>& testSpace, const EntityType& entity,
                      Dune::Stuff::LA::VectorInterface<V, R>& systemVector,
                      std::vector<std::vector<Dune::DynamicVector<R>>>& tmpLocalVectorContainer,
