+++ dune/gdt/basefunctionset/fem-localfunctions.hh
@@ -18,8 +18,8 @@ namespace BaseFunctionSet {
 
 
 // forward, to be used in the traits and to allow for specialization
-template <class BaseFunctionSetMapImp, class DomainFieldImp, int domainDim, class RangeFieldImp, int rangeDim,
-          int rangeDimCols = 1>
+template <class BaseFunctionSetMapImp, class DomainFieldImp, size_t domainDim, class RangeFieldImp, size_t rangeDim,
+          size_t rangeDimCols = 1>
 class FemLocalfunctionsWrapper
 {
   static_assert(AlwaysFalse<BaseFunctionSetMapImp>::value, "Untested for these dimensions!");
@@ -29,8 +29,8 @@ class FemLocalfunctionsWrapper
 namespace internal {
 
 
-template <class BaseFunctionSetMapImp, class DomainFieldImp, int domainDim, class RangeFieldImp, int rangeDim,
-          int rangeDimCols>
+template <class BaseFunctionSetMapImp, class DomainFieldImp, size_t domainDim, class RangeFieldImp, size_t rangeDim,
+          size_t rangeDimCols>
 class FemLocalfunctionsWrapperTraits
 {
 public:
@@ -44,7 +44,7 @@ public:
 } // namespace internal
 
 
-template <class BaseFunctionSetMapImp, class DomainFieldImp, int domainDim, class RangeFieldImp, int rangeDim>
+template <class BaseFunctionSetMapImp, class DomainFieldImp, size_t domainDim, class RangeFieldImp, size_t rangeDim>
 class FemLocalfunctionsWrapper<BaseFunctionSetMapImp, DomainFieldImp, domainDim, RangeFieldImp, rangeDim, 1>
     : public BaseFunctionSetInterface<internal::FemLocalfunctionsWrapperTraits<BaseFunctionSetMapImp, DomainFieldImp,
                                                                                domainDim, RangeFieldImp, rangeDim, 1>,
